refatoração, melhorando o código

+++ src/main/java/com/traineeprimeiro/domain/services/BaseService.java
@@ -1,12 +1,15 @@
 package com.traineeprimeiro.domain.services;
 
+import com.traineeprimeiro.domain.entities.EntityBase;
+import com.traineeprimeiro.domain.entities.cliente.Cliente;
 import com.traineeprimeiro.domain.entities.produto.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
 
-public class BaseService<T> {
+public class BaseService<T extends EntityBase> {
 
     protected JpaRepository<T, Long> _repository;
 
@@ -14,8 +17,9 @@ public class BaseService<T> {
         _repository = repository;
     }
 
-    public T Save(T entity) {
-        return this._repository.save(entity);
+    public boolean ExistsById(long id) {
+        boolean entityExists = this._repository.existsById(id);
+        return entityExists;
     }
 
     public List<T> FindAll() {
@@ -29,7 +33,22 @@ public class BaseService<T> {
         return entity.get();
     }
 
-    public void DeleteById(long id) {
-        this._repository.deleteById(id);
+    @Transactional
+    public T Save(T entity) {
+        entity.setId(0);
+        return this._repository.save(entity);
+    }
+
+    @Transactional
+    public T Update(T entity) {
+        boolean entityExists = this._repository.existsById(entity.getId());
+        if (entityExists) return null;
+        return this._repository.save(entity);
+    }
+
+    public boolean DeleteById(long id) {
+        boolean entityExists = this._repository.existsById(id);
+        if (entityExists) this._repository.deleteById(id);
+        return entityExists;
     }
 }
